Configure `sudo` access for an user toolbox
If setting up an user toolbox, add and configure `sudo` inside it, so
that the user can, for instance, install packages inside it.

The required packages ('sudo' and 'system-group-wheel') can either be
installed or added to the image. But even if we add them to the image,
considering that one can point toolbox to pulling any image, it probably
still make sense to try to install them.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>