Stay in current directory, even when entering a toolbox
In 99% of cases, one enters a user toolbox from somewhere under
his or hers own /home directory and would want/expect, once inside
the toolbox, to still be thera, rather thatn than in the directory
where he or she were when the toolbox was created.

In fact, this is not only quite convenient, but it also solves the
issue that the directory where we were back then may well not exist
any longer.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>