Connect to the host's SSH agent in user toolboxes.
Without this change, once inside a toolbox, ssh always asks for
the password, even if you have the proper credentials stored in
the host's SSH agent already, and this is rather annoying.

It is enough, however, to export to the container the environment
variable and the /run/user directory where the socket used by the
SSH agent lives, and ssh can be used happily from within the
toolbox as well.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>