Allow graphical applications from inside a user toolbox to run
It can happen that some user needs to quickly install and run
applications with a graphical interface, especially from user toolboxes.
That currently does not work, but making it working is just a matter of
exporting to the container a couple of env variables... So let's go for
it.

Another use case is when one use user toolboxes for some specific
tasks, such as development, testing or whatever. E.g., it would be quite
useful to build your application inside a toolbox (where you keep all
the dev and build tools and dependencies installed), and also be able to
test-run it, even if it is a graphical app.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>