Allow specifying the image to use via command line
Add `-R` parameter, for specifying the registry and `-I` parameter, for
specifying the image. If they're used, they override whatever it is foun
in the config file (or whatever the default is).

This can come in handy, for instance, if I want to create a development
toolbox that uses the image at registry.opensuse.org/opensuse/dario-dev-tbx
just once, while I'm fine with the default (or what's in config file)
for the other times. In such a case, it'd be quite annoying to have to
edit the config file itself for creating just that one toolbox (and then
edit it again for putting there the image I want for the most of the
times).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>