Randomize the name and cleanup the user setup script
In an user toolbox, we use a script to do some tweaking to the packages
and/or the user's group.

As reported in https://bugzilla.suse.com/show_bug.cgi?id=1178117 , it
having a predictable name may be an issue. Fix that by:
- giving it a random name
- putting it in the user home directory (indeed there are users without
  an home directory, but it's very unlikely that any of them tries to
  execute `toolbox -u`)

While there, let's also get rid of it, after it has done its job. We
don't need it to stay there anyway.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>