Make it easier to use custom images (#10)
* check if installing system-group-wheel is needed

The `system-group-wheel` package may be already included in the image.
If that is the case, there's no need to try installing it again!

This makes the creation of the toolbox faster, if that package is there
(and it currently is, in our default image).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>

* Allow specifying the image to use via command line

Add `-R` parameter, for specifying the registry and `-I` parameter, for
specifying the image. If they're used, they override whatever it is foun
in the config file (or whatever the default is).

This can come in handy, for instance, if I want to create a development
toolbox that uses the image at registry.opensuse.org/opensuse/dario-dev-tbx
just once, while I'm fine with the default (or what's in config file)
for the other times. In such a case, it'd be quite annoying to have to
edit the config file itself for creating just that one toolbox (and then
edit it again for putting there the image I want for the most of the
times).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>

* Randomize the name and cleanup the user setup script

In an user toolbox, we use a script to do some tweaking to the packages
and/or the user's group.

As reported in https://bugzilla.suse.com/show_bug.cgi?id=1178117 , it
having a predictable name may be an issue. Fix that by:
- giving it a random name
- putting it in the user home directory (indeed there are users without
  an home directory, but it's very unlikely that any of them tries to
  execute `toolbox -u`)

While there, let's also get rid of it, after it has done its job. We
don't need it to stay there anyway.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>