Introduce commands, for compatibility with Silverblue's toolbox
It has happened that users have tried to issue commands like
`toolbox create` or `toolbox enter`, which is how you interact with
toolbox on Fedora Silverblue or EndlessOS. And they're confused when
our toolbox does not work as they expect.

Let's, therefore, make our UI compatible with that. Of course, this is
without stopping supporting our original UI. If commands are used, user
toolbox are created and entered in, again because that is the kind of
toolbox an user coming from such OSes and using the command-based UI
expects.

So, basically, after this change, these are quivalent:
 * creating and entering a user toolbox:
   - toolbox -u
   - toolbox create
     toolbox enter
 * running /usr/bin/foo in an existing user toolbox:
   - toolbox -u /usr/bin/foo
   - toolbox run /usr/bin/foo
 * creating and entering in an user toolbox tagged -devel
   - toolbox -u -t devel
   - toolbox create -t devel
     toolbox enter -t devel

This will, hopefully, help users coming from other immutable OSes feel
"at home".

As a (nice) side effect, the following are also equivalent:
 * creating a user toolbox running as root:
   - toolbox -r -u
   - toolbox create -r
     toolbox enter -r
 * running /usr/bin/foo in an existing user toolbox running as root:
   - toolbox -r -u /usr/bin/foo
   - toolbox run -r /usr/bin/foo
 * creating and entering an user toolbox running as root:
   - toolbox -r -u -t devel
   - toolbox create -r -t devel
     toolbox enter -r -t devel

On the other hand, `toolbox list` is, for now, only an alias to
`podman ps` and `toolbox rm` and `rmi` (available in Silverblue) are
not implemented yet.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>