Be even more compatible with Silverblue Toolbox
Add a parameter for completely replacing with a given string (instead
than just adding a tag to) the container name of a toolbox. This is the
`-c` parameter and happens to be how that is done in Fedora Silverblue
toolbox too.

While there, also add `-i` as a way to specify a full URI for the image
to pull as a base of the toolbox. Basically, `-i reg.io/img` is
equivalent to `-R reg.io -I img`. This is again how this is done in the
other toolbox project.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>