Alleviate the need for zypper in the user toolbox script
Instead of (almost) always call zypper to install both sudo and
system-group-wheel, let's go for a simpler sudo setup, i.e., making the
toolbox user a sudoer, without going through %wheel.

This means that any image that has sudo installed can be used as
toolbox, even one of a distribution different than an [open]SUSE one,
that does not have zypper.

E.g., I was able to create a "fedora user toolbox", like this:

  $ toolbox enter -i docker.io/library/fedora -t fedora-test
  Spawning a container 'toolbox-dario-user-fedora-test' with image
  'docker.io/library/fedora'
  d2c4fcb317c25811954af5e17f7ae3a752b3287e1536d4621df16d85a1e283f5
  Container created.
  Entering container. To exit, type 'exit'.
  bash-5.0$ whoami
  dario
  bash-5.0$ sudo su
  [root@toolbox-dario-user-fedora-test microos-toolbox.git]# whoami
  root

If the image does not come with sudo, we still assume zypper and try to
install it. Ideally, we would want to make this general as well, or find
some other solution, but that will come as a future commit (once we'll
figure out the best way to do it).

Sicgned-off-by: Dario Faggioli <dfaggioli@suse.com>