Avoid copying the user setup script with 'podman cp'
We needed to copy the setup script (for user toolboxes), when they were
generated in /tmp on the host. But they're currently generated directly
in $HOME, and since we bind mount $HOME... well, it's just already
there (and one can even argue that it's a bug copying it!).

So, not copying is the right thing do to and, nicely enough, it also
workarounds what apparently is a podman 3 issue with 'podman cp',
basically fixing issue #25 too.

While there, add cleanup logic for the same user setup script. In fact,
if everything goes fine, the script is removed by the toolbox itself
(after executing it). But if there's an error and we bail, it may stick
around, cluttering the home directory.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>