Properly share namespaces in non-user toolboxes
When starting a "non-user" toolbox (i.e., starting without `-u`,
whether it's rootful or not), we currently do no share the PID and
IPC namespaces.

And that is bad, because non-user toolboxes are meant for debugging
and troubleshooting. And being able, say, to strace or attach GDB to
an host process is a super useful debugging feature.

And we need to share the host namespace, in order for that to be
possible.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>