Always pull when creating a new toolbox
Currently, if the image we need is available locally, we don't pull
it. However, the image we have locally might be months old, with the
result of creating an outdated toolbox, which is most likely not what
we want.

Let's, instead, always pull when creating, to be sure we start new
toolboxes fresh. Of course, this slows down new toolbox creation a
tiny bit, but it's worth it (and, also considering that toolbox is
stateful after all, creating new ones shouldn't be in any speed-critical
path).

Just in case, add a command line argument that lets one decide to skip
the pulling, if really wanting to.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>