Add a "more sandboxing" mode
toolbox is not meant for providing security and strong isolation, and
probably never will be. Still, users asked a simple way to start less
privileged toolboxes.

Of course, such a toolbox will be less convenient to use, because of the
lack of sharing with the host.

Of course^2, even in this mode, good security and strong isolation
should not be assumed!

https://github.com/kubic-project/microos-toolbox/issues/31

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>