Fix cleanup logic and make toolbox start a little less verbose
The temporary setup script is cleaned up right after finishing using
it, no need trying to clean it up again in cleanup()... Especially
considering that this does nothing anyway, as the name of the file
is in a local variable, not known by the cleanup trap!

Also avoid calling cleanup() twice in case of `toolbox enter` and
`toolbox run`.

While there, cleanup the terminal output a little bit, getting rid
of some annoying and not really useful messages

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>