Do not stop a toolbox with something running inside
If there's stuff running inside a toolbox (e.g., because it's still open
in another terminal, or something like that), let's not stop it.

So, this is the new semantic. When we exit a toolbox:
- if there is still something running inside, it keeps running;
- if there is nothing else running inside, and it was started _without_
  the `-n` flag, it is _stopped_;
- if there is nothing else running inside, and it was started *with*
  the  `-n` flag, it is *not* stopped.

In order to stop the containers that may have been left running, even if
they're "empty", we implement a `toolbox stop` command.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>