Exit if neither podman or docker are usable
Check early whether podman or docker are available and if they at
least seem to be working.

If that is not the case, there's no point going on.

While there, the list of variables to export to a user toolbox is
moved so that it is only evaluated when dealing with an actual
user toolbox, and in general closer to where it is useful.

This way, it does not show up all the time, even when it is not
useful (yet) when debugging (e.g., with `bash -x`).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>