Support passing just the name of the container to create and enter command
Currently, for specifying the container name, '-c <name>' must be used
(or we can add a tag to the default name with '-t <tag>'). It is however
much more convenient if, at least for the create and enter command
(i.e., only for user toolboxes!), we can just pass the name as an
argument, without the need of any option. E.g.:

  toolbox create work

Or:

  toolbox enter dev

Of course, using either '-c' or '-t' continues to be supported (although,
all these mechanisms are mutually exclusive, and only one must be used).

This is one more step toward CLI compatibility with
gh#containers/toolbox (and, to some extent, with distrobox too).

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>